--- db/migrate/20150911094916_create_productions.rb
@@ -1,8 +1,0 @@
-class CreateProductions < ActiveRecord::Migration
-  def change
-    create_table :productions do |t|
-      t.string      :title
-      t.timestamps  null: false
-    end
-  end
-end
