+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160124205556) do
+ActiveRecord::Schema.define(version: 20160225202000) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -29,7 +29,12 @@ ActiveRecord::Schema.define(version: 20160124205556) do
     t.string   "title"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.integer  "creator_id"
+    t.integer  "updater_id"
   end
+
+  add_index "productions", ["creator_id"], name: "index_productions_on_creator_id", using: :btree
+  add_index "productions", ["updater_id"], name: "index_productions_on_updater_id", using: :btree
 
   create_table "super_admins", id: false, force: :cascade do |t|
     t.integer  "user_id"
@@ -72,6 +77,8 @@ ActiveRecord::Schema.define(version: 20160124205556) do
 
   add_foreign_key "admins", "users"
   add_foreign_key "admins", "users", column: "assignor_id"
+  add_foreign_key "productions", "users", column: "creator_id"
+  add_foreign_key "productions", "users", column: "updater_id"
   add_foreign_key "super_admins", "users"
   add_foreign_key "suspensions", "users"
   add_foreign_key "suspensions", "users", column: "assignor_id"
