+++ spec/factories/production.rb
@@ -2,6 +2,8 @@ FactoryGirl.define do
 
   factory :production do
     title 'Hamlet'
+    association :creator, factory: :second_user
+    updater { creator }
 
     factory :add_production do
       title 'Othello'
